+++ rcfile/src/main/java/com/twitter/elephantbird/mapreduce/input/RCFileThriftInputFormat.java
@@ -6,17 +6,15 @@ import java.util.List;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.hive.ql.io.RCFileInputFormat;
 import org.apache.hadoop.hive.serde2.ColumnProjectionUtils;
 import org.apache.hadoop.hive.serde2.columnar.BytesRefArrayWritable;
 import org.apache.hadoop.hive.serde2.columnar.BytesRefWritable;
 import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.mapreduce.InputSplit;
 import org.apache.hadoop.mapreduce.RecordReader;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.hadoop.mapreduce.lib.input.FileSplit;
-import org.apache.pig.data.Tuple;
-import org.apache.pig.data.TupleFactory;
 import org.apache.thrift.TBase;
 import org.apache.thrift.TException;
 import org.apache.thrift.protocol.TBinaryProtocol;
@@ -27,26 +25,23 @@ import org.slf4j.LoggerFactory;
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
 import com.twitter.data.proto.Misc.ColumnarMetadata;
-import com.twitter.elephantbird.pig.util.RCFileUtil;
-import com.twitter.elephantbird.pig.util.ThriftToPig;
+import com.twitter.elephantbird.mapreduce.io.ThriftWritable;
+import com.twitter.elephantbird.util.RCFileUtil;
 import com.twitter.elephantbird.thrift.TStructDescriptor;
 import com.twitter.elephantbird.thrift.TStructDescriptor.Field;
 import com.twitter.elephantbird.util.ThriftUtils;
 import com.twitter.elephantbird.util.TypeRef;
 
-public class RCFileThriftInputFormat extends MapReduceInputFormatWrapper<LongWritable, BytesRefArrayWritable> {
+public class RCFileThriftInputFormat extends RCFileBaseInputFormat {
 
   private static final Logger LOG = LoggerFactory.getLogger(RCFileThriftInputFormat.class);
 
   private TypeRef<? extends TBase<?, ?>> typeRef;
 
-  /** internal, for MR use only. */
-  public RCFileThriftInputFormat() {
-    super(new RCFileInputFormat<LongWritable, BytesRefArrayWritable>());
-  }
+  // for MR
+  public RCFileThriftInputFormat() {}
 
   public RCFileThriftInputFormat(TypeRef<? extends TBase<?, ?>> typeRef) {
-    this();
     this.typeRef = typeRef;
   }
 
@@ -58,32 +53,34 @@ public class RCFileThriftInputFormat extends MapReduceInputFormatWrapper<LongWri
     ThriftUtils.setClassConf(conf, RCFileThriftInputFormat.class, thriftClass);
   }
 
-
-  @Override @SuppressWarnings("unchecked")
-  public RecordReader createRecordReader(InputSplit split,
-                                         TaskAttemptContext taskAttempt)
+  @Override
+  public RecordReader<LongWritable, Writable>
+  createRecordReader(InputSplit split, TaskAttemptContext taskAttempt)
                                     throws IOException, InterruptedException {
     if (typeRef == null) {
       typeRef = ThriftUtils.getTypeRef(taskAttempt.getConfiguration(), RCFileThriftInputFormat.class);
     }
-    return new ThriftReader(super.createRecordReader(split, taskAttempt));
+    return new ThriftReader(createUnwrappedRecordReader(split, taskAttempt));
   }
 
-  public class ThriftReader extends FilterRecordReader<LongWritable, BytesRefArrayWritable> {
+  public class ThriftReader extends FilterRecordReader<LongWritable, Writable> {
 
-    private final TupleFactory tf = TupleFactory.getInstance();
+    protected TStructDescriptor     tDesc;
+    protected boolean               readUnknownsColumn = false;
+    protected List<Field>           knownRequiredFields = Lists.newArrayList();
+    protected ArrayList<Integer>    columnsBeingRead = Lists.newArrayList();
 
-    private TStructDescriptor     tDesc;
-    private boolean               readUnknownsColumn = false;
-    private List<Field>           knownRequiredFields = Lists.newArrayList();
-    private ArrayList<Integer>    columnsBeingRead = Lists.newArrayList();
+    protected TMemoryInputTransport memTransport = new TMemoryInputTransport();
+    protected TBinaryProtocol tProto = new TBinaryProtocol(memTransport);
 
-    private TMemoryInputTransport memTransport = new TMemoryInputTransport();
-    private TBinaryProtocol tProto = new TBinaryProtocol(memTransport);
+    protected ThriftWritable<TBase<?, ?>> thriftWritable;
 
-    private TBase<?, ?>     currentValue;
-
-    public ThriftReader(RecordReader<LongWritable, BytesRefArrayWritable> reader) {
+    /**
+     * The reader is expected to be a
+     * <code>RecordReader< LongWritable, BytesRefArrayWritable ></code>
+     */
+    @SuppressWarnings({ "unchecked", "rawtypes" })
+    public ThriftReader(RecordReader reader) {
       super(reader);
     }
 
@@ -92,12 +89,13 @@ public class RCFileThriftInputFormat extends MapReduceInputFormatWrapper<LongWri
       return readUnknownsColumn;
     }
 
-    @Override
+    @Override @SuppressWarnings("unchecked")
     public void initialize(InputSplit split, TaskAttemptContext ctx)
                            throws IOException, InterruptedException {
       // set up columns that needs to read from the RCFile.
 
       tDesc = TStructDescriptor.getInstance(typeRef.getRawClass());
+      thriftWritable = ThriftWritable.newInstance((Class<TBase<?, ?>>)typeRef.getRawClass());
       final List<Field> tFields = tDesc.getFields();
 
       FileSplit fsplit = (FileSplit)split;
@@ -139,23 +137,31 @@ public class RCFileThriftInputFormat extends MapReduceInputFormatWrapper<LongWri
       super.initialize(split, ctx);
     }
 
-    @Override
-    public boolean nextKeyValue() throws IOException, InterruptedException {
-      currentValue = null;
-      return super.nextKeyValue();
+    @Override @SuppressWarnings({ "unchecked", "rawtypes" })
+    public Writable getCurrentValue() throws IOException, InterruptedException {
+      try {
+        thriftWritable.set(getCurrentThriftValue());
+        return thriftWritable;
+      } catch (TException e) {
+        //TODO : add error tracking
+        throw new IOException(e);
+      }
+    }
+
+    /** returns <code>super.getCurrentValue()</code> */
+    public BytesRefArrayWritable getCurrentBytesRefArrayWritable() throws IOException, InterruptedException {
+      return (BytesRefArrayWritable) super.getCurrentValue();
     }
 
     /**
      * Builds Thrift object from the raw bytes returned by RCFile reader.
      * @throws TException
      */
-    @SuppressWarnings("unchecked")
+    @SuppressWarnings({ "unchecked", "rawtypes" })
     public TBase<?, ?> getCurrentThriftValue() throws IOException, InterruptedException, TException {
-      if (currentValue != null) {
-        return currentValue;
-      }
 
-      BytesRefArrayWritable byteRefs = getCurrentValue();
+      BytesRefArrayWritable byteRefs = getCurrentBytesRefArrayWritable();
+
       if (byteRefs == null) {
         return null;
       }
@@ -185,39 +191,8 @@ public class RCFileThriftInputFormat extends MapReduceInputFormatWrapper<LongWri
         }
       }
 
-      currentValue = tObj;
-      return currentValue;
+      return tObj;
     }
 
-    /**
-     * Returns a Tuple consisting of required fields with out creating
-     * a Thrift message at the top level.
-     */
-    public Tuple getCurrentTupleValue() throws IOException, InterruptedException, TException {
-
-      BytesRefArrayWritable byteRefs = getCurrentValue();
-      if (byteRefs == null) {
-        return null;
-      }
-
-      Tuple tuple = tf.newTuple(knownRequiredFields.size());
-
-      for (int i=0; i < knownRequiredFields.size(); i++) {
-        BytesRefWritable buf = byteRefs.get(columnsBeingRead.get(i));
-        if (buf.getLength() > 0) {
-          memTransport.reset(buf.getData(), buf.getStart(), buf.getLength());
-
-          Field field = knownRequiredFields.get(i);
-          Object value = ThriftUtils.readFieldNoTag(tProto, field);
-          tuple.set(i, ThriftToPig.toPigObject(field, value, false));
-        }
-      }
-
-      if (readUnknownsColumn) {
-        throw new IOException("getCurrentTupleValue() is not supported when 'readUnknownColumns' is set");
-      }
-
-      return tuple;
-    }
   }
 }